+++ QuantumSim/sim/state.h
@@ -73,7 +73,7 @@ public:
                        const Gate& g = {});
     void ApplyHGateOnAllAmps(bool not_cycle_0 = false);
     void ApplyCGate(const int num_controls,
-                   const vector<int>& gate_qubits,
+                   const vector<size_t>& gate_qubits,
                    const Gate& g,
                    const Gate::Type gate_type);
     void ApplyMergedXYGate(const Gate& gate1,
