+++ QuantumSim/sim/state_tensor.cpp
@@ -417,7 +417,7 @@ ApplyHGateOnAllAmps(bool not_cycle_0)
 //TODO
 void TensorProductStateVector::
 ApplyCGate(const int num_controls,
-           const vector<int>& gate_qubits,
+           const vector<size_t>& gate_qubits,
            const Gate& g,
            const Gate::Type gate_type)
 {
